+++ SD_card.h
@@ -110,6 +110,7 @@ void Session_info(GPS_data G);
 void Session_results_M(GPS_speed M);
 void Session_results_S(GPS_time S);
 void Session_results_Alfa(Alfa_speed A,GPS_speed M);
+void Session_gpstc(char* gpstc);
 void TimeZone_env (float timezone);
 /*
 void log_header_SBP(void);